Refactor to separate files for readability
