+++ internal/ethereum/blocklistener.go
@@ -195,3 +195,7 @@ func (bl *blockListener) getHighestBlock() int64 {
 	bl.mux.Unlock()
 	return highestBlock
 }
+
+func (c *ethConnector) NewBlockHashes() <-chan *ffcapi.BlockHashEvent {
+	return nil
+}
